requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asn1crypto==1.5.1
base58==2.1.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coincurve==20.0.0
cytoolz==1.0.1
eth-hash==0.7.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastapi_pagination==0.12.33
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
parsimonious==0.10.0
psycopg2-binary==2.9.10
asyncpg==0.30.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.22
starlette==0.41.3
toolz==1.0.0
tronpy @ git+https://github.com/andelf/tronpy.git@18318b2a78851bd2bc6b9d2d870ddeacf7bce719
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
